Add Validation.Common.Job because there is too much duplicated code (#339)
Progress on https://github.com/NuGet/Engineering/issues/1168
