+++ src/NuGet.Services.Validation.Orchestrator/NuGet.Services.Validation.Orchestrator.csproj
@@ -99,80 +99,21 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Autofac">
-      <Version>4.6.1</Version>
-    </PackageReference>
-    <PackageReference Include="Autofac.Extensions.DependencyInjection">
-      <Version>4.2.0</Version>
-    </PackageReference>
-    <PackageReference Include="EntityFramework">
-      <Version>6.1.3</Version>
-    </PackageReference>
-    <PackageReference Include="Microsoft.ApplicationInsights">
-      <Version>2.2.0</Version>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.Configuration">
-      <Version>1.1.2</Version>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.Configuration.Binder">
-      <Version>1.1.2</Version>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.Configuration.FileExtensions">
-      <Version>1.1.2</Version>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection">
-      <Version>1.1.1</Version>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.Logging">
-      <Version>1.1.2</Version>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions">
-      <Version>1.1.2</Version>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions">
-      <Version>1.1.2</Version>
-    </PackageReference>
-    <PackageReference Include="NuGet.Services.Configuration">
-      <Version>2.14.0</Version>
-    </PackageReference>
-    <PackageReference Include="NuGet.Services.Contracts">
-      <Version>2.14.0</Version>
-    </PackageReference>
-    <PackageReference Include="NuGet.Services.KeyVault">
-      <Version>2.14.0</Version>
-    </PackageReference>
-    <PackageReference Include="NuGet.Services.Logging">
-      <Version>2.14.0</Version>
-    </PackageReference>
-    <PackageReference Include="NuGet.Services.ServiceBus">
-      <Version>2.14.0</Version>
-    </PackageReference>
-    <PackageReference Include="NuGet.Services.Validation">
-      <Version>2.14.0</Version>
-    </PackageReference>
     <PackageReference Include="NuGet.Services.Validation.Issues">
       <Version>2.14.0</Version>
     </PackageReference>
     <PackageReference Include="NuGet.Versioning">
       <Version>4.3.0</Version>
     </PackageReference>
-    <PackageReference Include="NuGetGallery.Core">
-      <Version>4.4.4-dev-19677</Version>
-    </PackageReference>
-    <PackageReference Include="Serilog">
-      <Version>2.5.0</Version>
-    </PackageReference>
-    <PackageReference Include="System.Net.Http">
-      <Version>4.3.3</Version>
-    </PackageReference>
-    <PackageReference Include="WindowsAzure.Storage">
-      <Version>7.1.2</Version>
-    </PackageReference>
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\NuGet.Jobs.Common\NuGet.Jobs.Common.csproj">
       <Project>{4b4b1efb-8f33-42e6-b79f-54e7f3293d31}</Project>
       <Name>NuGet.Jobs.Common</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\Validation.Common.Job\Validation.Common.Job.csproj">
+      <Project>{fa87d075-a934-4443-8d0b-5db32640b6d7}</Project>
+      <Name>Validation.Common.Job</Name>
     </ProjectReference>
     <ProjectReference Include="..\Validation.Common\Validation.Common.csproj">
       <Project>{2539ddf3-0cc5-4a03-b5f9-39b47744a7bd}</Project>
