+++ src/Validation.Common.Job/Error.cs
