+++ src/Validation.Common.Job/ExceptionExtensions.cs
