+++ src/Validation.Common.Job/Properties/AssemblyInfo.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("Validation.Common.Job")]
+[assembly: AssemblyDescription("Common logic for validation jobs.")]
+[assembly: ComVisible(false)]
+[assembly: Guid("fa87d075-a934-4443-8d0b-5db32640b6d7")]
