+++ src/Validation.Common.Job/Storage/AddStatusResult.cs
