+++ src/Validation.Common.Job/Storage/SaveStatusResult.cs
