+++ src/Validation.Common.Job/Storage/SerializedValidationIssue.cs
