+++ src/Validation.PackageSigning.Core/Validation.PackageSigning.Core.csproj
@@ -12,6 +12,7 @@
     <TargetFrameworkVersion>v4.6</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
+    <RestoreProjectStyle>PackageReference</RestoreProjectStyle>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>
@@ -38,38 +39,22 @@
     <Reference Include="System.Security" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="NuGet.Services.ServiceBus">
-      <Version>2.13.0</Version>
-    </PackageReference>
-    <PackageReference Include="NuGet.Services.Storage">
-      <Version>2.13.0</Version>
-    </PackageReference>
-    <PackageReference Include="NuGet.Services.Validation">
-      <Version>2.13.0</Version>
-    </PackageReference>
-    <PackageReference Include="WindowsAzure.Storage">
-      <Version>7.1.2</Version>
-    </PackageReference>
-  </ItemGroup>
-  <ItemGroup>
     <Compile Include="Configuration\CertificateStoreConfiguration.cs" />
-    <Compile Include="Error.cs" />
-    <Compile Include="ExceptionExtensions.cs" />
     <Compile Include="Extensions\CryptographicAttributeObjectCollectionExtensions.cs" />
     <Compile Include="Extensions\X509Certificate2Extensions.cs" />
-    <Compile Include="Storage\AddStatusResult.cs" />
     <Compile Include="Storage\CertificateStore.cs" />
     <Compile Include="Storage\ICertificateStore.cs" />
-    <Compile Include="Storage\IValidatorStateService.cs" />
     <Compile Include="Messages\CertificateValidationMessage.cs" />
     <Compile Include="Messages\CertificateValidationMessageSerializer.cs" />
     <Compile Include="Messages\SignatureValidationMessage.cs" />
     <Compile Include="Messages\SignatureValidationMessageSerializer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Storage\SaveStatusResult.cs" />
-    <Compile Include="Storage\SerializedValidationIssue.cs" />
-    <Compile Include="Storage\ValidatorStateService.cs" />
-    <Compile Include="Storage\ValidatorStatusExtensions.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Validation.Common.Job\Validation.Common.Job.csproj">
+      <Project>{fa87d075-a934-4443-8d0b-5db32640b6d7}</Project>
+      <Name>Validation.Common.Job</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
